Fix kernel DLC RX overflow while calculating CRCs. Update version to epoch. Size optimisations. Winforms: block app close while process running. Put close running kernel on tools menu. Print MD5 of loaded kernels to user log.

+++ Apps/PcmLibrary/CKernelWriter.cs
@@ -46,7 +46,7 @@ namespace PcmHacking
         /// </summary>
         public async Task<bool> Write(
             byte[] image,
-            UInt32 kernelVersion, 
+            UInt64 kernelVersion,
             FileValidator validator,
             bool needToCheckOperatingSystem,
             CancellationToken cancellationToken)
